+++ apps/demo/src/app/modules/doc/doc.ts
@@ -5,6 +5,7 @@ import { DOC_INTERACTION_ROOT_ROUTE } from './modules/interaction/doc.interactio
 import { DOC_LAYOUT_ROOT_ROUTE } from './modules/layout/doc.layout';
 import { DOC_ROUTER_ROOT_ROUTE } from './modules/router/doc.router';
 import { DOC_TEXT_ROOT_ROUTE } from './modules/text/doc.text';
+import { DOC_EXTENSION_ROOT_ROUTE } from './modules/extension/doc.extension';
 
 export const DOC_HOME_ROUTE = {
   icon: 'home',
@@ -12,7 +13,17 @@ export const DOC_HOME_ROUTE = {
   ref: 'doc.home'
 };
 
-export const DOC_ROUTES = [DOC_ACTION_ROOT_ROUTE, DOC_AUTH_ROOT_ROUTE, DOC_INTERACTION_ROOT_ROUTE, DOC_FORM_ROOT_ROUTE, DOC_LAYOUT_ROOT_ROUTE, DOC_ROUTER_ROOT_ROUTE, DOC_TEXT_ROOT_ROUTE];
+export const DOC_ROUTES = [
+  //
+  DOC_ACTION_ROOT_ROUTE,
+  DOC_AUTH_ROOT_ROUTE,
+  DOC_INTERACTION_ROOT_ROUTE,
+  DOC_FORM_ROOT_ROUTE,
+  DOC_EXTENSION_ROOT_ROUTE,
+  DOC_LAYOUT_ROOT_ROUTE,
+  DOC_ROUTER_ROOT_ROUTE,
+  DOC_TEXT_ROOT_ROUTE
+];
 
 export const DOC_ROOT_ROUTE = {
   icon: 'sensors',