+++ apps/demo/src/app/modules/doc/modules/extension/container/layout.component.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { ClickableAnchorLinkSegueRef } from '@dereekb/dbx-core';
+
+@Component({
+  templateUrl: './layout.component.html'
+})
+export class DocExtensionLayoutComponent {
+  readonly navAnchors: ClickableAnchorLinkSegueRef[] = [
+    {
+      title: 'Extensions',
+      ref: 'doc.extension'
+    }
+  ];
+}