+++ packages/date/src/lib/date/index.ts
@@ -1,3 +1,4 @@
+export * from './date.block';
 export * from './date.calendar';
 export * from './date.duration';
 export * from './date.format';