+++ packages/util/src/lib/date/date.ts
@@ -65,3 +65,4 @@ export type Days = number;
 export const MINUTES_IN_DAY = 1440;
 export const MS_IN_MINUTE = 1000 * 60;
 export const MS_IN_HOUR = MS_IN_MINUTE * 60;
+export const MS_IN_DAY = MS_IN_HOUR * 24;